requirements.txt
fire
nose2
pylint
ipykernel
PyYAML
ruamel.yaml
numpy
pandas
scipy
heartpy
loguru